fix: 터치패드 스크롤 관리 hook 추가 및 CardCarousel, ContentCarousel 반영
