+++ src/components/feature/Carousel/CardCarousel.module.css
@@ -2,7 +2,7 @@
     width: 100%;
     overflow-x: scroll;
     overflow-y: hidden;
-    scroll-behavior: smooth;
+    scroll-behavior: auto; /* smooth에서 auto로 변경 */
     -webkit-overflow-scrolling: touch;
     scrollbar-width: none; /* Firefox */
     -ms-overflow-style: none; /* IE and Edge */
